chore(js): inbox tabs moved tabs dropdown logic to hook
